+++ NEWS.md
@@ -3,6 +3,7 @@
 * Changed zero-fold per gene plot to show expression vs zero-fold. 
 * Added two static plots to About tab.
 * Added "top-correlated" panel to "Cell types" tab.
+* Added "inner-std" plot to QC panel. 
 
 # MCView 0.2.9